Changing the stub to be a single packaged assembly. This was accomplished by setting the IsPackable attribute to true in the csproj. The stub can only be used from projects with a new csproj due to the way libuv is resolved. Added in a way to specify ports on startup.cs. Added some brief docs to the readme and some xml comments on all public memebers.
