--- UKHO.ConfigurableStub/UKHO.ConfigurableStub.LocalDevelopmentConsoleApp/Program.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-
-namespace UKHO.ConfigurableStub.LocalDevelopmentConsoleApp
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var StubManager = new StubManager();
-            StubManager.Start();
-            Console.ReadLine();
-        }
-    }
-}
