--- UKHO.ConfigurableStub/UKHO.ConfigurableStub.LocalDevelopmentConsoleApp/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net461" />
-</packages>
