+++ UKHO.ConfigurableStub/UKHO.ConfigurableStub.sln
@@ -3,11 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.329
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UKHO.ConfigurableStub", "UKHO.ConfigurableStub\UKHO.ConfigurableStub.csproj", "{80A7467A-81F9-4374-A6C5-5FE776F82560}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UKHO.ConfigurableStub.Stub", "UKHO.ConfigurableStub.Stub\UKHO.ConfigurableStub.Stub.csproj", "{B9B7D0FE-A7B5-418B-BDCC-5639CFE8F3F8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UKHO.ConfigurableStub.LocalDevelopmentConsoleApp", "UKHO.ConfigurableStub.LocalDevelopmentConsoleApp\UKHO.ConfigurableStub.LocalDevelopmentConsoleApp.csproj", "{1A27454D-BC5A-479E-B607-439647D6BF65}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,18 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{80A7467A-81F9-4374-A6C5-5FE776F82560}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{80A7467A-81F9-4374-A6C5-5FE776F82560}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{80A7467A-81F9-4374-A6C5-5FE776F82560}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{80A7467A-81F9-4374-A6C5-5FE776F82560}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B9B7D0FE-A7B5-418B-BDCC-5639CFE8F3F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B9B7D0FE-A7B5-418B-BDCC-5639CFE8F3F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B9B7D0FE-A7B5-418B-BDCC-5639CFE8F3F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B9B7D0FE-A7B5-418B-BDCC-5639CFE8F3F8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1A27454D-BC5A-479E-B607-439647D6BF65}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1A27454D-BC5A-479E-B607-439647D6BF65}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1A27454D-BC5A-479E-B607-439647D6BF65}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1A27454D-BC5A-479E-B607-439647D6BF65}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
