--- UKHO.ConfigurableStub/UKHO.ConfigurableStub/UKHO.ConfigurableStub.csproj
@@ -1,38 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net461</TargetFramework>
-    <TargetsForTfmSpecificBuildOutput>$(TargetsForTfmSpecificBuildOutput);IncludeStubAssets</TargetsForTfmSpecificBuildOutput>
-    <TargetsForTfmSpecificBuildOutput>$(TargetsForTfmSpecificBuildOutput);IncludeStubAssets</TargetsForTfmSpecificBuildOutput>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <DebugSymbols>true</DebugSymbols>
-    <ApplicationIcon />
-    <PlatformTarget>x64</PlatformTarget>
-    <OutputType>Library</OutputType>
-    <StartupObject />
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
-    <PackageReference Include="System.Net.Http" Version="4.3.3" />
-  </ItemGroup>
-
-
-  <ItemGroup>
-    <ProjectReference Include="..\UKHO.ConfigurableStub.Stub\UKHO.ConfigurableStub.Stub.csproj" PrivateAssets="All">
-    </ProjectReference>
-  </ItemGroup>
-
-
-
-  <Target Name="IncludeStubAssets">
-    <ItemGroup>
-      <BuildOutputInPackage Include="$(OutputPath)*.*" />
-    </ItemGroup>
-  </Target>
-  <Target Name="IncludeStubAssets">
-    <ItemGroup>
-      <BuildOutputInPackage Include="$(OutputPath)*.*" />
-    </ItemGroup>
-  </Target>
-</Project>
